Deliberately forbid blocking redis commands
Note that SUPPORT_BLOCKING_REDIS_COMMAND_UNSAFE is disabled by default,
and would have to be manually enabled with `CFLAGS='-DSUPPORT_BLOCKING_REDIS_COMMAND_UNSAFE=1'`
during compilation

It's way too easy to cause severe and unpredictable performance degradation for other clients
if an application unexpectedly starts uses blocking redis commands,
and it'd be much easier to not support this command to begin with.
(twemproxy uses pipelining, so the blocking command would block other requests, possibly indefinitely).

Additionally, the configured timeout may be hit if the command timeout is less than the twemproxy timeout.