Improve heartbeat, failover, sentinel, and nutcracker stability
Fix memory leaks in the heartbeat patches and mark hosts as failed
before regenerating the heartbeat continuum.

Use a consistent definition of a failed host for the heartbeat patches
to avoid creating a pool with 0 hosts.

Document the heartbeat patches.

Fix a segmentation fault caused by using the number of points in a
continuum instead of the number of servers - the points in a continuum
can be 0 if all hosts are ejected at once.

Optimization: If a memcache/redis multiget has only one key, forward it
directly instead of splitting it up into fragmented commands.